+++ src/main/java/com/kinoticket/backend/security/WebSecurityConfig.java
@@ -1,5 +1,7 @@
 package com.kinoticket.backend.security;
 
+import com.kinoticket.backend.service.UserService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -23,7 +25,7 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 )
 public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
     @Autowired
-    private UserDetailsServiceImpl userDetailsService;
+    private UserService userDetailsService;
 
     @Autowired
     private AuthEntryPointJwt unauthorizedHandler;
