+++ src/main/java/com/kinoticket/backend/service/UserService.java
@@ -1,10 +1,11 @@
-package com.kinoticket.backend.security;
+package com.kinoticket.backend.service;
 
 import javax.transaction.Transactional;
 
 import com.kinoticket.backend.model.User;
 import com.kinoticket.backend.repositories.UserRepository;
 import com.kinoticket.backend.repositories.VerificationTokenRepository;
+import com.kinoticket.backend.security.VerificationToken;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -13,7 +14,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UserDetailsServiceImpl implements UserDetailsService {
+public class UserService implements UserDetailsService {
 
     @Autowired
     UserRepository userRepository;
@@ -27,7 +28,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
         User user = userRepository.findByUsername(username)
             .orElseThrow(() -> new UsernameNotFoundException("User Not Found with username: " + username));
 
-        return UserDetailsImpl.build(user);
+        return user;
     }
 
     public void createVerificationToken(User user, String token) {
